src/containers/green-screen-matting/model/requirements.txt
base64image==0.5.1
numpy==1.22.0
Pillow==9.1.1
onnxruntime-gpu==1.8.1
boto3
flask
gevent